requirements.txt
# 데이터 처리
pandas==2.0.3
numpy==1.24.3

# API 요청
requests==2.31.0

# 머신러닝
scikit-learn==1.3.0

# 웹 인터페이스
streamlit==1.28.0

# 자연어 처리 (선택)
konlpy==0.6.0

# 시각화
matplotlib==3.7.2
seaborn==0.12.2

# 제미나이 
google-generativeai>=0.3.1